feat(core): money columns FLOAT -> Numeric(18,4)
Every monetary column (unit_cost, extended_cost, prices, PO/line totals,
make-vs-buy and should-cost breakdowns, vendor/contract pricing) was
stored as Float, so repeated money arithmetic silently drifted by
fractions of a cent (binary floats cannot represent decimals like 0.1
exactly). Switches all of them to Numeric(18,4), matching the precision
already used elsewhere in this schema (bom.py, inventory.py, revision.py,
routing.py, labor.py). Non-monetary Float columns (scores, ratings,
weights, percentages, hours, days, grams) are left untouched.

Adds chained alembic migration 033 (single head confirmed) and fixes two
Decimal/float arithmetic TypeErrors this surfaced in the make-vs-buy and
should-cost update endpoints, where partially-updated fields (fresh
Decimal from the DB) were mixed with untouched-field floats from the
request payload.

TDD: app/tests/test_money_precision.py persists Decimal(0.1)+Decimal(0.2)
and repeated-add totals across Part/BomItem/POLineItem/MakeVsBuyAnalysis/
ShouldCostModel/PartVendor, round-trips through a fresh session, and
asserts exact equality -- fails on Float, passes on Numeric(18,4).

+++ backend/app/models/bom_item.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, DateTime, Float, ForeignKey, Index, Integer, String, Text
+from sqlalchemy import Column, DateTime, ForeignKey, Index, Integer, Numeric, String, Text
 from sqlalchemy.orm import relationship
 from sqlalchemy.sql import func
 
@@ -25,8 +25,8 @@ class BomItem(Base, TenantAwareMixin):
     )  # For sub-BOM hierarchy
 
     # Cost snapshot at time of BOM creation
-    unitCostSnapshot = Column(Float)
-    extendedCost = Column(Float)
+    unitCostSnapshot = Column(Numeric(18, 4))
+    extendedCost = Column(Numeric(18, 4))
 
     # Timestamps
     createdAt = Column(DateTime(timezone=True), server_default=func.now())
